ios details

+++ ios/Runner/Model/DiscoverMoviesModels.swift
@@ -29,7 +29,7 @@ struct Result: Codable {
     let posterPath, releaseDate, title: String
     let video: Bool
     let voteAverage: Double
-    let voteCount: Double
+    let voteCount: Int
 
     enum CodingKeys: String, CodingKey {
         case adult
